Add files via upload
Merit Award winners of the Singapore Space Challenge-2021 (here on will be known as the International Space Challenge). 

We were given 3 months to design an entire lunar rover along with full operations and mission plan. We designed a lunar rover capable of performing a topographical study of the Lunar Lava Tubes present on the moon using an onboard GPR and a deep-learning-based RTAB navigation system. 

Our entry was shortlisted in the top 10 proposals from around the world and we won the Merit Award out of 139 Global teams from 18 countries.